+ change out app.xaml to use dynamicresource instead of directly using solarized:Theme. -> allows us to change styles at will + Add style for TextBlock / TextBox + Add handler in App.xaml.cs to update styles when the theme changes + Change HexContent's constants from ints to doubles so we can use them in Xaml. + MainWindow: add shortcuts/menuitems for Goto, Back, and Forward. + MainWindow: add control for Goto + MainWindow: add control/logic for toggling the Theme + MainWindow: add control to show addresses + EditorViewModel: add commands for back, forward, goto, and showGoto -> UI shortcut for Goto can't just immediately do a Goto: more info is needed.    so instead, the first UI command just shows the box (and moves the focus)    and a second command performs the action. + ScrollRegion: implement Goto/Back/Forward + ScrollRegion: own the logic for updating the address headers
This is the initial push for the goto feature. There are bugs / needs more tests.

+++ Gen3Hex/View/HexContent.cs
@@ -11,7 +11,7 @@ using System.Windows.Media;
 
 namespace HavenSoft.Gen3Hex.View {
    public class HexContent : FrameworkElement {
-      public const int CellWidth = 30, CellHeight = 20;
+      public const double CellWidth = 30, CellHeight = 20;
 
       public static readonly Rect CellRect = new Rect(0, 0, CellWidth, CellHeight);
 
@@ -152,13 +152,13 @@ namespace HavenSoft.Gen3Hex.View {
       }
 
       private void UpdateViewPortSize() {
-         ViewPort.Width = (int)ActualWidth / CellWidth;
-         ViewPort.Height = (int)ActualHeight / CellHeight;
+         ViewPort.Width = (int)(ActualWidth / CellWidth);
+         ViewPort.Height = (int)(ActualHeight / CellHeight);
       }
 
       private Model.Point ControlCoordinatesToModelCoordinates(MouseEventArgs e) {
          var point = e.GetPosition(this);
-         return new Model.Point((int)point.X / CellWidth, (int)point.Y / CellHeight);
+         return new Model.Point((int)(point.X / CellWidth), (int)(point.Y / CellHeight));
       }
    }
 }
